refactor: #385: move position table to entities
