+++ src/entities/position/index.ts
@@ -1,0 +1,2 @@
+export { Positions } from './ui/table';
+export type { PositionsTableProps } from './ui/table';
