+++ src/entities/position/ui/dash.tsx
