+++ src/entities/position/ui/error-notice.tsx
