+++ src/entities/position/ui/no-positions.tsx
