+++ src/pages/trade/ui/history-tabs.tsx
@@ -4,7 +4,8 @@ import { Tabs } from '@penumbra-zone/ui/Tabs';
 import { Density } from '@penumbra-zone/ui/Density';
 import { Toggle } from '@penumbra-zone/ui/Toggle';
 import { Text } from '@penumbra-zone/ui/Text';
-import Positions from './positions';
+import { Positions } from '@/entities/position';
+import { usePathToMetadata } from '../model/use-path';
 
 enum PositionsTabsType {
   MY_POSITIONS = 'MY_POSITIONS',
@@ -14,6 +15,7 @@ export const HistoryTabs = () => {
   const [parent] = useAutoAnimate();
   const [tab, setTab] = useState<PositionsTabsType>(PositionsTabsType.MY_POSITIONS);
   const [showInactive, setshowInactive] = useState(false);
+  const { baseAsset, quoteAsset } = usePathToMetadata();
 
   return (
     <div ref={parent} className='flex flex-col w-screen desktop:w-auto'>
@@ -33,7 +35,7 @@ export const HistoryTabs = () => {
         </label>
       </div>
 
-      <Positions showInactive={showInactive} />
+      <Positions showInactive={showInactive} base={baseAsset} quote={quoteAsset} />
     </div>
   );
 };
